requirements.txt
Pillow==9.4.0
requests==2.28.2
python-dotenv==0.21.1